Add updatePermissions

+++ src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const JWT = require('jsonwebtoken');
 const { randomBytes} = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
     async createItem(parent, args, ctx, info) {
@@ -39,8 +40,13 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
         //1. find the item 
-        const item = await ctx.db.query.item({ where }, `{ id title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user {id} }`);
         //2. check the permissions
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+        if (!ownsItem && !hasPermissions) {
+            throw new Error('You dont have permissions to do that');
+        } 
         //3. delete it 
         return ctx.db.mutation.deleteItem({ where }, info);
     },
@@ -175,8 +181,34 @@ const Mutations = {
         })
         //return the new user 
         return updateUser;
-
+    },
+    async updatePermissions(parent, args, ctx, info) {
+        //check if the are logged in 
+        if (!ctx.request.userId) {
+            throw new Error('You must be logged in');
+        }
+        //query the current user
+        const currentUser = await ctx.db.query.user({
+            where: {
+                id: ctx.request.userId
+            }
+        }, info);
+        //check if they have permission to do this
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        //update the permissions
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    //because permissions is enum, we use set
+                    set: args.permissions
+                }
+            },
+            where: {
+                id: args.userId
+            }
+        }, info);
     }
+
  };
 
 module.exports = Mutations;
